complete basic network communication.

+++ src/server/shared/Packets/WorldPacket.h
@@ -30,7 +30,7 @@ class WorldPacket : public ByteBuffer
         {
         }
 
-        explicit WorldPacket(uint16 opcode, size_t res=200) : ByteBuffer(res), m_opcode(opcode) { }
+        explicit WorldPacket(uint32 opcode, size_t res=200) : ByteBuffer(res), m_opcode(opcode) { }
 
         WorldPacket(WorldPacket&& packet) : ByteBuffer(std::move(packet)), m_opcode(packet.m_opcode)
         {
@@ -51,20 +51,20 @@ class WorldPacket : public ByteBuffer
             return *this;
         }
 
-        WorldPacket(uint16 opcode, MessageBuffer&& buffer) : ByteBuffer(std::move(buffer)), m_opcode(opcode) { }
+        WorldPacket(uint32 opcode, MessageBuffer&& buffer) : ByteBuffer(std::move(buffer)), m_opcode(opcode) { }
 
-        void Initialize(uint16 opcode, size_t newres=200)
+        void Initialize(uint32 opcode, size_t newres=200)
         {
             clear();
             _storage.reserve(newres);
             m_opcode = opcode;
         }
 
-        uint16 GetOpcode() const { return m_opcode; }
-        void SetOpcode(uint16 opcode) { m_opcode = opcode; }
+        uint32 GetOpcode() const { return m_opcode; }
+        void SetOpcode(uint32 opcode) { m_opcode = opcode; }
 
     protected:
-        uint16 m_opcode;
+        uint32 m_opcode;
 };
 
 #endif
